removed unused, renamed some files, added inCombat / inCreation validation
