+++ frontend/src/components/Creation/Options.js
@@ -40,7 +40,7 @@ const Options = ({ classes }) => {
             onChange={handleOpen}
           >
           {[...monsters].sort((a, b) => a.name.localeCompare(b.name))
-          .map(monster => <MenuItem key={monster.name} value={monster.name}>{monster.name}</MenuItem>)}
+          .map(monster => <MenuItem disabled={monster.inCombat} key={monster.name} value={monster.name}>{monster.name}</MenuItem>)}
         </Select>
       </FormControl>}
     </div>
