--- frontend/src/state/actions/locationActions.ts
@@ -1,6 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export const getSelectedTabAction = props => ({
-  type: actionTypes.GET_SELECTED_TAB,
-  payload: props
-})
